page: Add support for options
Instead of taking only a title option, turn the first argument into a
map, and support other options like the URIs to the CSS, impress.js
and of course the page title.

If the first argument is not a map, then impress will call itself with
a dummy options map, so that the defaults get used.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>